// app/src/generated/models/Csa_customersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_customersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_customersstatecode = keyof typeof Csa_customersstatecode;
export const Csa_customersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_customersstatuscode = keyof typeof Csa_customersstatuscode;

export interface Csa_customersBase {
  csa_company?: string;
  csa_customerid: string;
  csa_email?: string;
  csa_lastinteractiondate?: string;
  csa_name: string;
  csa_role?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_customersstatecode;
  statuscode?: Csa_customersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_customers extends Csa_customersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}